requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
mysqlclient==2.2.7
packaging==26.0
pillow==12.1.0
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
